+++ sembast/lib/src/database_impl.dart
@@ -119,8 +119,11 @@ class SembastDatabase extends Object
   SembastStore get mainStore => _mainStore;
 
   SembastStore _mainStore;
-  final Map<String, SembastStore> _stores = {};
-  final List<String> _txnDroppedStores = [];
+  final _stores = <String, SembastStore>{};
+  final _txnDroppedStores = <String>[];
+
+  /// To optimize auto increment int key generation in a transaction
+  final _txnStoreLastIntKeys = <String, int>{};
 
   /// Current in memory stores
   Iterable<SembastStore> get stores => _stores.values;
@@ -139,6 +142,7 @@ class SembastDatabase extends Object
 
   void _clearTxnData() {
     _txnDroppedStores.clear();
+    _txnStoreLastIntKeys.clear();
 
     // remove temp data in all store
     for (var store in stores) {
@@ -232,9 +236,20 @@ class SembastDatabase extends Object
       store.currentRecords;
 
   /// For jdb only
-  Future<int> generateUniqueIntKey(String store) {
+  Future<int> generateUniqueIntKey(String store) async {
     if (_storageJdb != null) {
-      return _storageJdb.generateUniqueIntKey(store);
+      // Get any previous generated int key
+      var lastIntKey = _txnStoreLastIntKeys[store];
+      if (lastIntKey == null) {
+        lastIntKey = await _storageJdb.generateUniqueIntKey(store);
+      } else {
+        // increment previous read value
+        lastIntKey++;
+      }
+      // Save for laters insert in the same transaction
+      _txnStoreLastIntKeys[store] = lastIntKey;
+
+      return lastIntKey;
     }
     return null;
   }
@@ -1132,12 +1147,16 @@ class SembastDatabase extends Object
                 var entry = JdbWriteEntry()..txnRecord = record;
                 entries.add(entry);
               }
+              final infoEntries = <JdbInfoEntry>[
+                if (upgrading) getMetaInfoEntry(commitEntries.upgradingMeta)
+              ];
+              _txnStoreLastIntKeys.forEach((store, lastId) {
+                infoEntries.add(getStoreLastIntKeyInfoEntry(store, lastId));
+              });
               var query = StorageJdbWriteQuery(
                   revision: commitEntries.revision,
                   entries: entries,
-                  infoEntries: upgrading
-                      ? [getMetaInfoEntry(commitEntries.upgradingMeta)]
-                      : null);
+                  infoEntries: infoEntries);
 
               /// Commit to storage now
               var status = await storageJdb.writeIfRevision(query);
@@ -1268,19 +1287,6 @@ class SembastDatabase extends Object
           FutureOr<T> Function(SembastTransaction transaction) action) =>
       transaction((txn) => action(txn as SembastTransaction));
 
-  /// records must not changed
-  Future forEachRecords(List<ImmutableSembastRecord> records,
-      void Function(ImmutableSembastRecord record) action) async {
-    // handle record in transaction first
-    for (var record in records) {
-      // then the regular unless already in transaction
-      if (needCooperate) {
-        await cooperate();
-      }
-      action(record);
-    }
-  }
-
   // Only set during open
   @override
   SembastTransaction get sembastTransaction =>
@@ -1300,40 +1306,6 @@ class SembastDatabase extends Object
           if (debugListener) {
             print('full import listeners none');
           }
-          /*
-          // records, synchronous
-          var store = operation.store;
-          var keys = operation.listener.recordKeys;
-          for (var key in keys) {
-            var record = store.record(key);
-            var ctlrs = operation.listener.getRecordControllers(record);
-            for (var ctlr in ctlrs) {
-              ctlr.add(operation.content.record(key));
-            }
-          }
-          // store, async
-          // Fix existing queries
-          for (var query in List<QueryListenerController>.from(
-              operation.listener.getQuery())) {
-            Future _updateQuery() async {
-              if (debugListener) {
-                print('updating $query: with ${operation.content.records}');
-              }
-              await query.update(operation.content.records, cooperator);
-            }
-
-            if (query.hasInitialData) {
-              await _updateQuery();
-            } else {
-              // postpone
-              // ignore: unawaited_futures
-              notificationLock.synchronized(() async {
-                await _updateQuery();
-              });
-            }
-          }
-           */
-
         } else {
           for (var record in operation.content.records) {
             var ctlrs = operation.listener.getRecordControllers(record.ref);